requirements.txt
beautifulsoup4==4.11.1
nltk==3.7
pandas==1.4.2
tweet-preprocessor==0.6.0
requests==2.27.1
scikit_learn==1.1.1
tweet_preprocessor==0.6.0
numpy==1.22.4
plotly==5.8.0